+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed:
 - Show open, groomed & in-progress tasks by default
+- Remove AppBar in Journal
 
 ## [0.8.20] - 2022-05-28
 ### Fixed: